chore(DATAHUB-37): migrate interfaces into central file
