+++ src/assets/content.ts
@@ -1,24 +1,22 @@
-interface Project {
-  title: string;
-  city: string;
-  description: string;
-}
-// type ProjectNameTypes = "temp_grunewald" | "testabc";
+import { ProjectType } from "../common/interfaces";
+
 interface Content {
   projects: {
-    [key: string]: Project;
+    [key: string]: ProjectType;
   };
 }
 
 export const content: Content = {
   projects: {
     temp_grunewald: {
+      id: "temp_grunewald",
       title: "TSB-Temperatur",
       city: "Berlin, Deutschland",
       description:
         "Hier steht eine Beschreibung des Projekts. Hier steht eine Beschreibung.",
     },
     testabc: {
+      id: "testabc",
       title: "Test-Projekt",
       city: "Berlin, Deutschland",
       description:
