+++ src/components/ProjectPreview.tsx
@@ -2,7 +2,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import { jsx, Box, Card, Heading, Text, Grid } from "theme-ui";
-import { ProjectType } from "../state/model";
+import { ProjectType } from "../common/interfaces";
 
 export const ProjectPreview: React.FC<ProjectType> = ({
   id,
