+++ src/state/model.ts
@@ -1,22 +1,5 @@
 import { Thunk, Action, Computed } from "easy-peasy";
-
-export interface DeviceType {
-  description: string;
-  id: number;
-  latitude: number;
-  longitude: number;
-  ttnAppId: string;
-  ttnDeviceId: string;
-  records?: Array<any>;
-}
-
-export interface ProjectType {
-  id: string;
-  title: string;
-  city: string;
-  description: string;
-  devices: Array<DeviceType>;
-}
+import { ProjectType, DeviceType } from "../common/interfaces";
 
 export interface ProjectsModel {
   items: Array<ProjectType> | undefined;
